refactor: group all router related things in the http package
